# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TourWorld API
  description: |
    TourWorld 여행 관리 시스템 API

    ## 주요 기능
    - **일정표 관리 (FEAT-2)**: 여행 일정표 생성, 편집, 관리
    - **호텔 관리 (FEAT-1)**: 호텔 정보 등록 및 관리
    - **이미지 갤러리 (FEAT-3)**: 이미지 업로드 및 카테고리 관리

    ## 인증
    JWT 토큰 기반 인증을 사용합니다.
    - `Authorization: Bearer <token>` 헤더를 통해 인증

    ## 에러 응답
    모든 에러는 동일한 형식으로 반환됩니다:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "에러 메시지",
        "details": {}
      }
    }
    ```
  version: 1.0.0
  contact:
    name: TourWorld Team
servers:
  - url: http://localhost:3000/api
    description: 개발 서버
  - url: https://api.tourworld.com/api
    description: 프로덕션 서버
...
paths:
...
  /itineraries:
    get:
      tags:
        - Itineraries
      summary: 일정표 목록 조회
      operationId: listItineraries
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ItineraryStatus'
        - name: search
          in: query
          schema:
            type: string
          description: 제목, 설명 검색
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 일정표 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListItinerariesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Itineraries
      summary: 일정표 생성
      operationId: createItinerary
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItineraryRequest'
      responses:
        '201':
          description: 일정표 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateItineraryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /itineraries/{id}:
    get:
      tags:
        - Itineraries
      summary: 일정표 상세 조회
      operationId: getItinerary
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 일정표 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetItineraryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Itineraries
      summary: 일정표 수정
      operationId: updateItinerary
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItineraryRequest'
      responses:
        '200':
          description: 일정표 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateItineraryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Itineraries
      summary: 일정표 삭제
      operationId: deleteItinerary
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 일정표 삭제됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # Itinerary Items (FEAT-2)
  # ==========================================
...
  /images/categories:
    get:
      tags:
        - ImageCategories
      summary: 이미지 카테고리 목록 조회
      operationId: listImageCategories
      security:
        - bearerAuth: []
      parameters:
        - name: includeSystem
          in: query
          schema:
            type: boolean
            default: true
          description: 시스템 기본 카테고리 포함 여부
      responses:
        '200':
          description: 이미지 카테고리 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListImageCategoriesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ImageCategories
      summary: 이미지 카테고리 생성
      operationId: createImageCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageCategoryRequest'
      responses:
        '201':
          description: 카테고리 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateImageCategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /images/categories/{id}:
    get:
      tags:
        - ImageCategories
      summary: 이미지 카테고리 상세 조회
      operationId: getImageCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 카테고리 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetImageCategoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - ImageCategories
      summary: 이미지 카테고리 수정
      operationId: updateImageCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateImageCategoryRequest'
      responses:
        '200':
          description: 카테고리 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateImageCategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - ImageCategories
      summary: 이미지 카테고리 삭제
      operationId: deleteImageCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 카테고리 삭제됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 인증 토큰

  # ==========================================
  # Parameters
  # ==========================================
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 리소스 ID
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 페이지 번호
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 페이지당 항목 수
    SortByParam:
      name: sortBy
      in: query
      schema:
        type: string
      description: 정렬 기준 필드
    SortOrderParam:
      name: sortOrder
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: 정렬 순서

  # ==========================================
  # Responses
  # ==========================================
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: 입력값 검증에 실패했습니다.
              details:
                errors:
                  - field: title
                    message: 필수 필드입니다.
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: 인증이 필요합니다.
...
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 요청한 리소스를 찾을 수 없습니다.
    Conflict:
      description: 리소스 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: CONFLICT
              message: 리소스 충돌이 발생했습니다.
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        field:
          type: string
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalCount
        - totalPages
        - hasNextPage
        - hasPreviousPage
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
...
    DeleteResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          properties:
            deleted:
              type: boolean
              enum: [true]
...
    ItineraryStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ARCHIVED]
    ItineraryMetadata:
      type: object
      properties:
        templateId:
          type: string
        branding:
          type: object
          properties:
            primaryColor:
              type: string
            logoUrl:
              type: string
            companyName:
              type: string
        settings:
          type: object
          properties:
            showMap:
              type: boolean
            showWeather:
              type: boolean
            showTimezone:
              type: boolean
    Itinerary:
      type: object
      required:
        - id
        - userId
        - title
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/ItineraryStatus'
        metadata:
          $ref: '#/components/schemas/ItineraryMetadata'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ItinerarySummary:
      type: object
      required:
        - id
        - title
        - status
        - itemCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/ItineraryStatus'
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateItineraryRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/ItineraryStatus'
        metadata:
          $ref: '#/components/schemas/ItineraryMetadata'
    UpdateItineraryRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/ItineraryStatus'
        metadata:
          $ref: '#/components/schemas/ItineraryMetadata'
    ListItinerariesResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required:
            - items
            - meta
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ItinerarySummary'
            meta:
              $ref: '#/components/schemas/PaginationMeta'
    GetItineraryResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/Itinerary'
    CreateItineraryResponse:
      $ref: '#/components/schemas/GetItineraryResponse'
    UpdateItineraryResponse:
      $ref: '#/components/schemas/GetItineraryResponse'

    # ItineraryItem Schemas
...
    ImageCategory:
      type: object
      required:
        - id
        - name
        - sortOrder
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
          description: null이면 시스템 기본 카테고리
        name:
          type: string
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    ImageCategoryWithCount:
      allOf:
        - $ref: '#/components/schemas/ImageCategory'
        - type: object
          required:
            - imageCount
          properties:
            imageCount:
              type: integer
    CreateImageCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
    UpdateImageCategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
        sortOrder:
          type: integer
    ListImageCategoriesResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/ImageCategoryWithCount'
    GetImageCategoryResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/ImageCategory'
    CreateImageCategoryResponse:
      $ref: '#/components/schemas/GetImageCategoryResponse'
    UpdateImageCategoryResponse:
      $ref: '#/components/schemas/GetImageCategoryResponse'